+++ dbt/adapters/glue/column.py
@@ -1,11 +1,11 @@
 from dataclasses import dataclass
 from typing import ClassVar, Dict
 
-from dbt.adapters.base.column import Column
+from dbt.adapters.spark.column import SparkColumn
 
 
 @dataclass
-class GlueColumn(Column):
+class GlueColumn(SparkColumn):
     # Overwriting dafult string types to support glue
     # TODO: Convert to supported glue types as needed
     # Please ref: https://github.com/dbt-athena/dbt-athena/blob/main/dbt/adapters/athena/column.py